backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = "⚡ Ship RAG Solutions Quickly"
authors = [ "Adithya S K <Adithyaskolavi@gmail.com>" ]
readme = "README.md"

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = {extras = ["all"], version = "^0.112.2"}
python-dotenv = "^1.0.0"
aiostream = "^0.5.2"
llama-index = "0.10.58"
cachetools = "^5.3.3"
pymongo = "^4.8.0"
motor = "^3.5.1"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
fastapi-mail = "^1.4.1"
beanie = "^1.26.0"
python-decouple = "^3.8"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.llama-index-vector-stores-qdrant]
version = "^0.2.8"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[tool.poetry.dependencies.llama-index-agent-openai]
version = "0.2.6"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"